+++ src/compas_timber/connections/plate_butt_joint.py
@@ -93,7 +93,7 @@ class PlateButtJoint(PlateJoint):
         for polyline in self.main_outlines:
             move_polyline_segment_to_plane(polyline, self.main_segment_index, self.cross_planes[0])
 
-        if self.topology == JointTopology.TOPO_L:
+        if self.topology == JointTopology.TOPO_EDGE_EDGE:
             for polyline in self.cross_outlines:
                 move_polyline_segment_to_plane(polyline, self.cross_segment_index, self.main_planes[1])
 
@@ -121,7 +121,7 @@ class PlateButtJoint(PlateJoint):
 class PlateLButtJoint(PlateButtJoint):
     """Creates a plate-to-plate butt-joint connection."""
 
-    SUPPORTED_TOPOLOGY = JointTopology.TOPO_L
+    SUPPORTED_TOPOLOGY = JointTopology.TOPO_EDGE_EDGE
 
     def __repr__(self):
         return "PlateLButtJoint({0}, {1}, {2})".format(self.main_plate, self.cross_plate, JointTopology.get_name(self.topology))
@@ -142,7 +142,7 @@ class PlateLButtJoint(PlateButtJoint):
 class PlateTButtJoint(PlateButtJoint):
     """Creates a plate-to-plate butt-joint connection."""
 
-    SUPPORTED_TOPOLOGY = JointTopology.TOPO_T
+    SUPPORTED_TOPOLOGY = JointTopology.TOPO_EDGE_FACE
 
     def __repr__(self):
         return "PlateTButtJoint({0}, {1}, {2})".format(self.main_plate, self.cross_plate, JointTopology.get_name(self.topology))